fixed according to comments. Added AbstractEs6_4ClientInstrumentationTest
