+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/test/java/co/elastic/apm/agent/es/restclient/v6_4/AbstractEs6_4ClientInstrumentationTest.java
@@ -1,0 +1,250 @@
+package co.elastic.apm.agent.es.restclient.v6_4;
+
+import co.elastic.apm.agent.es.restclient.AbstractEsClientInstrumentationTest;
+import co.elastic.apm.agent.impl.transaction.Span;
+import org.elasticsearch.ElasticsearchStatusException;
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
+import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
+import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
+import org.elasticsearch.action.bulk.BulkRequest;
+import org.elasticsearch.action.bulk.BulkResponse;
+import org.elasticsearch.action.delete.DeleteRequest;
+import org.elasticsearch.action.delete.DeleteResponse;
+import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.action.index.IndexResponse;
+import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.action.search.SearchResponse;
+import org.elasticsearch.action.support.WriteRequest;
+import org.elasticsearch.action.support.master.AcknowledgedResponse;
+import org.elasticsearch.action.update.UpdateRequest;
+import org.elasticsearch.action.update.UpdateResponse;
+import org.elasticsearch.client.RequestOptions;
+import org.elasticsearch.client.RestHighLevelClient;
+import org.elasticsearch.index.query.QueryBuilders;
+import org.elasticsearch.search.SearchHits;
+import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.junit.Test;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
+
+public class AbstractEs6_4ClientInstrumentationTest extends AbstractEsClientInstrumentationTest {
+
+    protected static final String USER_NAME = "elastic-user";
+    protected static final String PASSWORD = "elastic-pass";
+
+    @SuppressWarnings("NullableProblems")
+    protected static RestHighLevelClient client;
+
+    @Test
+    public void testCreateAndDeleteIndex() throws IOException, ExecutionException, InterruptedException {
+        // Create an Index
+        doCreateIndex(new CreateIndexRequest(SECOND_INDEX));
+
+        validateSpanContentAfterIndexCreateRequest();
+        // Delete the index
+        reporter.reset();
+
+        doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
+
+        validateSpanContentAfterIndexDeleteRequest();
+    }
+
+    @Test
+    public void testTryToDeleteNonExistingIndex() throws IOException, InterruptedException {
+        ElasticsearchStatusException ese = null;
+        try {
+            doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
+        } catch (ElasticsearchStatusException e) {
+            // sync scenario
+            ese = e;
+        } catch (ExecutionException e) {
+            // async scenario
+            ese = (ElasticsearchStatusException) e.getCause();
+        }
+        assertThat(ese).isNotNull();
+        assertThat(ese.status().getStatus()).isEqualTo(404);
+
+        assertThatErrorsExistWhenDeleteNonExistingIndex();
+    }
+
+    @Test
+    public void testDocumentScenario() throws Exception {
+        // Index a document
+        IndexResponse ir = doIndex(new IndexRequest(INDEX, DOC_TYPE, DOC_ID).source(
+            jsonBuilder()
+                .startObject()
+                .field(FOO, BAR)
+                .endObject()
+        ).setRefreshPolicy(WriteRequest.RefreshPolicy.IMMEDIATE));
+        assertThat(ir.status().getStatus()).isEqualTo(201);
+
+        System.out.println(reporter.generateTransactionPayloadJson());
+
+        List<Span> spans = reporter.getSpans();
+        assertThat(spans).hasSize(1);
+        validateSpanContent(spans.get(0), String.format("Elasticsearch: PUT /%s/%s/%s", INDEX, DOC_TYPE, DOC_ID), 201, "PUT");
+
+        // Search the index
+        reporter.reset();
+
+        SearchRequest searchRequest = new SearchRequest(INDEX);
+        SearchSourceBuilder sourceBuilder = new SearchSourceBuilder();
+        sourceBuilder.query(QueryBuilders.termQuery(FOO, BAR));
+        sourceBuilder.from(0);
+        sourceBuilder.size(5);
+        searchRequest.source(sourceBuilder);
+        SearchResponse sr = doSearch(searchRequest);
+        verifyTotalHits(sr.getHits());
+        System.out.println(reporter.generateTransactionPayloadJson());
+
+        spans = reporter.getSpans();
+        assertThat(spans).hasSize(1);
+        Span searchSpan = spans.get(0);
+        validateSpanContent(searchSpan, String.format("Elasticsearch: POST /%s/_search", INDEX), 200, "POST");
+        validateDbContextContent(searchSpan, "{\"from\":0,\"size\":5,\"query\":{\"term\":{\"foo\":{\"value\":\"bar\",\"boost\":1.0}}}}");
+
+        // Now update and re-search
+        reporter.reset();
+
+        Map<String, Object> jsonMap = new HashMap<>();
+        jsonMap.put(FOO, BAZ);
+        UpdateRequest updateRequest = new UpdateRequest(INDEX, DOC_TYPE, DOC_ID).doc(jsonMap).setRefreshPolicy(WriteRequest.RefreshPolicy.IMMEDIATE);
+        UpdateResponse ur = doUpdate(updateRequest);
+        assertThat(ur.status().getStatus()).isEqualTo(200);
+        sr = doSearch(new SearchRequest(INDEX));
+        assertThat(sr.getHits().getAt(0).getSourceAsMap().get(FOO)).isEqualTo(BAZ);
+
+        System.out.println(reporter.generateTransactionPayloadJson());
+
+        spans = reporter.getSpans();
+        assertThat(spans).hasSize(2);
+        boolean updateSpanFound = false;
+        for(Span span: spans) {
+            if(span.getName().toString().contains("_update")) {
+                updateSpanFound = true;
+                break;
+            }
+        }
+        assertThat(updateSpanFound).isTrue();
+
+        // Finally - delete the document
+        reporter.reset();
+        DeleteResponse dr = doDelete(new DeleteRequest(INDEX, DOC_TYPE, DOC_ID));
+        assertThat(dr.status().getStatus()).isEqualTo(200);
+        validateSpanContent(spans.get(0), String.format("Elasticsearch: DELETE /%s/%s/%s", INDEX, DOC_TYPE, DOC_ID), 200, "DELETE");
+
+        System.out.println(reporter.generateTransactionPayloadJson());
+    }
+
+    protected void verifyTotalHits(SearchHits searchHits) {
+
+    }
+
+    @Test
+    public void testScenarioAsBulkRequest() throws IOException, ExecutionException, InterruptedException {
+        doBulk(new BulkRequest()
+            .add(new IndexRequest(INDEX, DOC_TYPE, "2").source(
+                jsonBuilder()
+                    .startObject()
+                    .field(FOO, BAR)
+                    .endObject()
+            ))
+            .add(new DeleteRequest(INDEX, DOC_TYPE, "2")));
+
+        validateSpanContentAfterBulkRequest();
+    }
+
+
+    private interface ClientMethod<Req, Res> {
+        void invoke(Req request, RequestOptions options, ActionListener<Res> listener);
+    }
+
+    private <Req, Res> Res invokeAsync(Req request, ClientMethod<Req, Res> method) throws InterruptedException, ExecutionException {
+        final CompletableFuture<Res> resultFuture = new CompletableFuture<>();
+        method.invoke(request, RequestOptions.DEFAULT, new ActionListener<>() {
+            @Override
+            public void onResponse(Res response) {
+                resultFuture.complete(response);
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                resultFuture.completeExceptionally(e);
+            }
+        });
+        return resultFuture.get();
+    }
+
+    protected CreateIndexResponse doCreateIndex(CreateIndexRequest createIndexRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<CreateIndexRequest, CreateIndexResponse> method =
+                (request, options, listener) -> client.indices().createAsync(request, options, listener);
+            return invokeAsync(createIndexRequest, method);
+        }
+        return client.indices().create(createIndexRequest, RequestOptions.DEFAULT);
+    }
+
+    protected AcknowledgedResponse doDeleteIndex(DeleteIndexRequest deleteIndexRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<DeleteIndexRequest, AcknowledgedResponse> method =
+                (request, options, listener) -> client.indices().deleteAsync(request, options, listener);
+            return invokeAsync(deleteIndexRequest, method);
+        }
+        return client.indices().delete(deleteIndexRequest, RequestOptions.DEFAULT);
+    }
+
+    protected IndexResponse doIndex(IndexRequest indexRequest) throws ExecutionException, InterruptedException, IOException {
+        if (async) {
+            ClientMethod<IndexRequest, IndexResponse> method =
+                (request, options, listener) -> client.indexAsync(request, options, listener);
+            return invokeAsync(indexRequest, method);
+        }
+        return client.index(indexRequest, RequestOptions.DEFAULT);
+    }
+
+    protected SearchResponse doSearch(SearchRequest searchRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<SearchRequest, SearchResponse> method =
+                (request, options, listener) -> client.searchAsync(request, options, listener);
+            return invokeAsync(searchRequest, method);
+        }
+        return client.search(searchRequest, RequestOptions.DEFAULT);
+    }
+
+    protected UpdateResponse doUpdate(UpdateRequest updateRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<UpdateRequest, UpdateResponse> method =
+                (request, options, listener) -> client.updateAsync(request, options, listener);
+            return invokeAsync(updateRequest, method);
+        }
+        return client.update(updateRequest, RequestOptions.DEFAULT);
+    }
+
+    protected DeleteResponse doDelete(DeleteRequest deleteRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<DeleteRequest, DeleteResponse> method =
+                (request, options, listener) -> client.deleteAsync(request, options, listener);
+            return invokeAsync(deleteRequest, method);
+        }
+        return client.delete(deleteRequest, RequestOptions.DEFAULT);
+    }
+
+    protected BulkResponse doBulk(BulkRequest bulkRequest) throws IOException, ExecutionException, InterruptedException {
+        if (async) {
+            ClientMethod<BulkRequest, BulkResponse> method =
+                (request, options, listener) -> client.bulkAsync(request, options, listener);
+            return invokeAsync(bulkRequest, method);
+        }
+        return client.bulk(bulkRequest, RequestOptions.DEFAULT);
+    }
+
+}
