+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/test/java/co/elastic/apm/agent/es/restclient/v6_4/ElasticsearchRestClientInstrumentationIT.java
@@ -24,58 +24,33 @@
  */
 package co.elastic.apm.agent.es.restclient.v6_4;
 
-import co.elastic.apm.agent.es.restclient.AbstractEsClientInstrumentationTest;
-import co.elastic.apm.agent.impl.transaction.Span;
 import org.apache.http.HttpHost;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
 import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.elasticsearch.ElasticsearchStatusException;
-import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
-import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
-import org.elasticsearch.action.bulk.BulkRequest;
-import org.elasticsearch.action.bulk.BulkResponse;
-import org.elasticsearch.action.delete.DeleteRequest;
-import org.elasticsearch.action.delete.DeleteResponse;
-import org.elasticsearch.action.index.IndexRequest;
-import org.elasticsearch.action.index.IndexResponse;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.support.WriteRequest;
-import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.action.update.UpdateResponse;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestClient;
 import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.elasticsearch.search.SearchHits;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.testcontainers.elasticsearch.ElasticsearchContainer;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
+
 
 @RunWith(Parameterized.class)
-public class ElasticsearchRestClientInstrumentationIT extends AbstractEsClientInstrumentationTest {
+public class ElasticsearchRestClientInstrumentationIT extends AbstractEs6_4ClientInstrumentationTest {
 
-    @SuppressWarnings("NullableProblems")
-    private static RestHighLevelClient client;
+    private static final String ELASTICSEARCH_CONTAINER_VERSION = "docker.elastic.co/elasticsearch/elasticsearch:6.7.0";
 
     public ElasticsearchRestClientInstrumentationIT(boolean async) {
         this.async = async;
@@ -84,7 +59,7 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractEsClientIn
     @BeforeClass
     public static void startElasticsearchContainerAndClient() throws IOException {
         // Start the container
-        container = new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:6.7.0");
+        container = new ElasticsearchContainer(ELASTICSEARCH_CONTAINER_VERSION);
         container.start();
 
         // Create the client
@@ -106,203 +81,10 @@ public class ElasticsearchRestClientInstrumentationIT extends AbstractEsClientIn
         client.close();
     }
 
-    @Test
-    public void testTryToDeleteNonExistingIndex() throws IOException, InterruptedException {
-        ElasticsearchStatusException ese = null;
-        try {
-            doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
-        } catch (ElasticsearchStatusException e) {
-            // sync scenario
-            ese = e;
-        } catch (ExecutionException e) {
-            // async scenario
-            ese = (ElasticsearchStatusException) e.getCause();
-        }
-        assertThat(ese).isNotNull();
-        assertThat(ese.status().getStatus()).isEqualTo(404);
-
-        assertThatErrorsExistWhenDeleteNonExistingIndex();
+    @Override
+    protected void verifyTotalHits(SearchHits searchHits) {
+        assertThat(searchHits.totalHits).isEqualTo(1L);
+        assertThat(searchHits.getAt(0).getSourceAsMap().get(FOO)).isEqualTo(BAR);
     }
 
-    @Test
-    public void testCreateAndDeleteIndex() throws IOException, ExecutionException, InterruptedException {
-        // Create an Index
-        doCreateIndex(new CreateIndexRequest(SECOND_INDEX));
-
-        validateSpanContentAfterIndexCreateRequest();
-        // Delete the index
-        reporter.reset();
-
-        doDeleteIndex(new DeleteIndexRequest(SECOND_INDEX));
-
-        validateSpanContentAfterIndexDeleteRequest();
-    }
-
-    @Test
-    public void testDocumentScenario() throws Exception {
-        // Index a document
-        IndexResponse ir = doIndex(new IndexRequest(INDEX, DOC_TYPE, DOC_ID).source(
-            jsonBuilder()
-                .startObject()
-                .field(FOO, BAR)
-                .endObject()
-        ).setRefreshPolicy(WriteRequest.RefreshPolicy.IMMEDIATE));
-        assertThat(ir.status().getStatus()).isEqualTo(201);
-
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        List<Span> spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: PUT /%s/%s/%s", INDEX, DOC_TYPE, DOC_ID), 201, "PUT");
-
-        // Search the index
-        reporter.reset();
-
-        SearchRequest searchRequest = new SearchRequest(INDEX);
-        SearchSourceBuilder sourceBuilder = new SearchSourceBuilder();
-        sourceBuilder.query(QueryBuilders.termQuery(FOO, BAR));
-        sourceBuilder.from(0);
-        sourceBuilder.size(5);
-        searchRequest.source(sourceBuilder);
-        SearchResponse sr = doSearch(searchRequest);
-        assertThat(sr.getHits().totalHits).isEqualTo(1L);
-        assertThat(sr.getHits().getAt(0).getSourceAsMap().get(FOO)).isEqualTo(BAR);
-
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        spans = reporter.getSpans();
-        assertThat(spans).hasSize(1);
-        Span searchSpan = spans.get(0);
-        validateSpanContent(searchSpan, String.format("Elasticsearch: POST /%s/_search", INDEX), 200, "POST");
-        validateDbContextContent(searchSpan, "{\"from\":0,\"size\":5,\"query\":{\"term\":{\"foo\":{\"value\":\"bar\",\"boost\":1.0}}}}");
-
-        // Now update and re-search
-        reporter.reset();
-
-        Map<String, Object> jsonMap = new HashMap<>();
-        jsonMap.put(FOO, BAZ);
-        UpdateRequest updateRequest = new UpdateRequest(INDEX, DOC_TYPE, DOC_ID).doc(jsonMap).setRefreshPolicy(WriteRequest.RefreshPolicy.IMMEDIATE);
-        UpdateResponse ur = doUpdate(updateRequest);
-        assertThat(ur.status().getStatus()).isEqualTo(200);
-        sr = doSearch(new SearchRequest(INDEX));
-        assertThat(sr.getHits().getAt(0).getSourceAsMap().get(FOO)).isEqualTo(BAZ);
-
-        System.out.println(reporter.generateTransactionPayloadJson());
-
-        spans = reporter.getSpans();
-        assertThat(spans).hasSize(2);
-        boolean updateSpanFound = false;
-        for(Span span: spans) {
-            if(span.getName().toString().contains("_update")) {
-                updateSpanFound = true;
-                break;
-            }
-        }
-        assertThat(updateSpanFound).isTrue();
-
-        // Finally - delete the document
-        reporter.reset();
-        DeleteResponse dr = doDelete(new DeleteRequest(INDEX, DOC_TYPE, DOC_ID));
-        assertThat(dr.status().getStatus()).isEqualTo(200);
-        validateSpanContent(spans.get(0), String.format("Elasticsearch: DELETE /%s/%s/%s", INDEX, DOC_TYPE, DOC_ID), 200, "DELETE");
-
-        System.out.println(reporter.generateTransactionPayloadJson());
-    }
-
-    @Test
-    public void testScenarioAsBulkRequest() throws IOException, ExecutionException, InterruptedException {
-        doBulk(new BulkRequest()
-            .add(new IndexRequest(INDEX, DOC_TYPE, "2").source(
-                jsonBuilder()
-                    .startObject()
-                    .field(FOO, BAR)
-                    .endObject()
-            ))
-            .add(new DeleteRequest(INDEX, DOC_TYPE, "2")));
-
-        validateSpanContentAfterBulkRequest();
-    }
-
-    private interface ClientMethod<Req, Res> {
-        void invoke(Req request, RequestOptions options, ActionListener<Res> listener);
-    }
-
-    private <Req, Res> Res invokeAsync(Req request, ClientMethod<Req, Res> method) throws InterruptedException, ExecutionException {
-        final CompletableFuture<Res> resultFuture = new CompletableFuture<>();
-        method.invoke(request, RequestOptions.DEFAULT, new ActionListener<>() {
-            @Override
-            public void onResponse(Res response) {
-                resultFuture.complete(response);
-            }
-
-            @Override
-            public void onFailure(Exception e) {
-                resultFuture.completeExceptionally(e);
-            }
-        });
-        return resultFuture.get();
-    }
-
-    private CreateIndexResponse doCreateIndex(CreateIndexRequest createIndexRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<CreateIndexRequest, CreateIndexResponse> method =
-                (request, options, listener) -> client.indices().createAsync(request, options, listener);
-            return invokeAsync(createIndexRequest, method);
-        }
-        return client.indices().create(createIndexRequest, RequestOptions.DEFAULT);
-    }
-
-    private AcknowledgedResponse doDeleteIndex(DeleteIndexRequest deleteIndexRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<DeleteIndexRequest, AcknowledgedResponse> method =
-                (request, options, listener) -> client.indices().deleteAsync(request, options, listener);
-            return invokeAsync(deleteIndexRequest, method);
-        }
-        return client.indices().delete(deleteIndexRequest, RequestOptions.DEFAULT);
-    }
-
-    private IndexResponse doIndex(IndexRequest indexRequest) throws ExecutionException, InterruptedException, IOException {
-        if (async) {
-            ClientMethod<IndexRequest, IndexResponse> method =
-                (request, options, listener) -> client.indexAsync(request, options, listener);
-            return invokeAsync(indexRequest, method);
-        }
-        return client.index(indexRequest, RequestOptions.DEFAULT);
-    }
-
-    private SearchResponse doSearch(SearchRequest searchRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<SearchRequest, SearchResponse> method =
-                (request, options, listener) -> client.searchAsync(request, options, listener);
-            return invokeAsync(searchRequest, method);
-        }
-        return client.search(searchRequest, RequestOptions.DEFAULT);
-    }
-
-    private UpdateResponse doUpdate(UpdateRequest updateRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<UpdateRequest, UpdateResponse> method =
-                (request, options, listener) -> client.updateAsync(request, options, listener);
-            return invokeAsync(updateRequest, method);
-        }
-        return client.update(updateRequest, RequestOptions.DEFAULT);
-    }
-
-    private DeleteResponse doDelete(DeleteRequest deleteRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<DeleteRequest, DeleteResponse> method =
-                (request, options, listener) -> client.deleteAsync(request, options, listener);
-            return invokeAsync(deleteRequest, method);
-        }
-        return client.delete(deleteRequest, RequestOptions.DEFAULT);
-    }
-
-    private BulkResponse doBulk(BulkRequest bulkRequest) throws IOException, ExecutionException, InterruptedException {
-        if (async) {
-            ClientMethod<BulkRequest, BulkResponse> method =
-                (request, options, listener) -> client.bulkAsync(request, options, listener);
-            return invokeAsync(bulkRequest, method);
-        }
-        return client.bulk(bulkRequest, RequestOptions.DEFAULT);
-    }
 }
